kira/perception/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kira-perception"
version = "0.1.0"
description = "Kira perception services - VLM, STT, and TTS for visual AI companion"
requires-python = ">=3.10"
dependencies = [
    # Core
    "numpy>=1.24",
    "pillow>=10.0",
    # VLM - Moondream
    "torch>=2.0",
    "transformers>=4.36",
    "accelerate>=0.25",
    "einops>=0.7",
    "pyvips>=2.2",  # Required by moondream2
    # STT - Whisper
    "openai-whisper>=20231117",
    "sounddevice>=0.4",
    # TTS - Chatterbox
    "chatterbox-tts>=0.1",
    "torchaudio>=2.0",
    "onnx>=1.15,<1.18",  # Pin to avoid ml_dtypes compatibility issue
    # Camera (for VLM tests with real camera)
    "opencv-python>=4.8",
    "torchvision>=0.21.0",
    "silero-vad>=6.2.0",
    "faster-whisper>=1.2.1",
    "piper-tts>=1.3.0",
    "deepface>=0.0.96",
    "tf-keras>=2.15.1",
    "mediapipe>=0.10.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "integration: marks tests as integration tests (require models/hardware)",
    "slow: marks tests as slow (model loading)",
]

[tool.hatch.build.targets.wheel]
packages = ["vlm", "audio", "tts"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
